fix p347 compile error

+++ solved/p347/p347.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-    int i, j, k, *p;
+    long i, j, k, *p;
     int **table;
     int *ntable;
     long tmp;
